Add 'locking' mechanism for source faces and embeddings (Alt+click_
Use Alt+click to toggle "lock" a source face or saved embedding to always be on and highlighted in green. When locked, source faces and saved embeddings won't be cleared when selecting new target faces or input images/videos, even with auto-swap toggled off. This allows the end user to maintain a specific "look" for a target face, even between videos.