Set absolute location of timeline handle
This ensures that when switching media during playback, the handle is still correctly placed. Previously, the handle was just moved by a delta each update based on the previous frame number, but if playback continues when media is changed, this can result in the delta being huge or even negative, outside the bounds of the timeline. This method ensures the handle is always within the timeline widget's bounds.