Add another face canvas for selected faces
These can be weighted individually with the mouse wheel, reverted to default weighting with a left click, or removed from the embedding with right click.

The face weight can also be typed directly into the entry box under each selected face.

Defaults for each 'index' face can be defined in the parameters panel, in case you always start with a specific face or merge and blend it into other faces.

-Separate clear_faces into clear_target_faces and clear_selected_source_faces
-create local functions for multiple code paths in  add_selected_face_to_selected_faces_canvas to be more easily shared